// src/app/[lang]/_common/types/session.type.ts
import { ISODateString } from "next-auth";

enum Gender {
  MALE = "Male",
  FEMALE = "Female",
}
export enum UserRole {
  ADMIN = "admin",
  "company-occupant" = "company-occupant",
  "organization-owner" = "organization-owner",
}
export type CustomSession = {
  user: {
    token: string;
    personalInformation: {
      firstName: string;
      lastName: string;
      gender: Gender;
      dateOfBirth: string;
    };
    contactInformation: {
      address: string;
      phone: string;
      email: string;
    };
    profileInformation?: {
      picture: string;
    };
    preferences?: {
      language: string;
      theme: string;
    };
    permissions: { role: UserRole|string; organization: "string" };
  };
  expires?: ISODateString;
};